fixes #30 deleting same cursor position
The issue was with the CodeMirrorAdapter, not actually anything wrong with the operational transform logic. The issue was that the editor was ignoring change events from CodeMirror if the change was applied by the other party, so that you don't keep sending the same change back and forth. However if both participants delete the same position then the change from the other participant becomes a no-op and the editor does not trigger a change event. The result is that the subsequent change event is ignored instead which results in the editor going out of sync.